# Fix localisation overhead in `SpriteText` / `LocalisedBindableString`