master:

**benchmark will not complete** (performance consistently degrades enough from the `WeakList` overhead that benchmark.net will not allow it).

### with only `UnbindAll` commit (https://github.com/ppy/osu-framework/commit/a2ff29a183b4173de8229cd41521d2651a8f8fa2)

|                Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| BenchmarkNonLocalised | 827.4 ns |  9.59 ns |  8.01 ns | 0.0591 |     - |     - |     680 B |
|    BenchmarkLocalised | 939.9 ns | 14.54 ns | 11.35 ns | 0.0668 |     - |     - |     776 B |


### also with lazy binding commit (https://github.com/ppy/osu-framework/commit/59a4b8920359143a226a457849f4c3921e635959):

|                Method |      Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |----------:|----------:|----------:|-------:|------:|------:|----------:|
| BenchmarkNonLocalised |  68.50 ns |  1.397 ns |  1.307 ns | 0.0131 |     - |     - |     152 B |
|    BenchmarkLocalised | 919.32 ns | 18.094 ns | 16.925 ns | 0.0677 |     - |     - |     784 B |

the extra 8 bytes here is the reference to `LocalisationManager`.

Closes #4679.

I *think* the benchmark not completing on master is enough by way of testing, but we could add proper testing by exposing some internals of `WeakList` if thats required.